ethereum-developer-bootcamp/final-project/ erc20-token-balance-tracker/requirements.txt
argparse==1.4.0
pandas==1.5.3
pycoingecko==3.1.0
python-dotenv==0.21.1
schedule==1.1.0
tabulate==0.9.0
